+++ src/LanguageServer/Microsoft.CodeAnalysis.LanguageServer.UnitTests/VSCodeSettingsTests.cs
@@ -53,42 +53,6 @@ public sealed class VSCodeSettingsTests : IDisposable
         Assert.Equal("disable", settings.TryGetStringSetting(VSCodeSettings.Names.DefaultSolution));
     }
 
-    [Fact]
-    public void TryGetVSCodeSolutionSettings_SingleFolderDisableSuppressesAutoLoad()
-    {
-        var workspaceFolder = CreateWorkspaceFolder("single", """
-            {
-              "dotnet.defaultSolution": "disable"
-            }
-            """);
-
-        var (isLoadingDisabled, solutionPath) = AutoLoadProjectsInitializer.TryGetVSCodeSolutionSettings([workspaceFolder], NullLogger.Instance);
-
-        Assert.True(isLoadingDisabled);
-        Assert.Null(solutionPath);
-    }
-
-    [Fact]
-    public void TryGetVSCodeSolutionSettings_MultiFolderUsesFirstConfiguredSolution()
-    {
-        var firstFolder = CreateWorkspaceFolder("first", """
-            {
-              "dotnet.defaultSolution": "disable"
-            }
-            """);
-        var secondFolder = CreateWorkspaceFolder("second", """
-            {
-              "dotnet.defaultSolution": "eng/App.slnx"
-            }
-            """);
-        CreateSolutionFile(secondFolder, "eng/App.slnx");
-
-        var (isLoadingDisabled, solutionPath) = AutoLoadProjectsInitializer.TryGetVSCodeSolutionSettings([firstFolder, secondFolder], NullLogger.Instance);
-
-        Assert.False(isLoadingDisabled);
-        Assert.Equal(Path.Combine(ProtocolConversions.GetDocumentFilePathFromUri(secondFolder.DocumentUri.GetRequiredParsedUri()), "eng", "App.slnx"), solutionPath);
-    }
-
     private string WriteSettingsFile(string settingsJson)
     {
         var folder = _tempRoot.CreateDirectory();
@@ -103,25 +67,5 @@ public sealed class VSCodeSettingsTests : IDisposable
         var settingsPath = WriteSettingsFile(settingsJson);
         Assert.True(VSCodeSettings.TryRead(settingsPath, NullLogger.Instance, out var settings));
         return settings;
-    }
-
-    private WorkspaceFolder CreateWorkspaceFolder(string name, string settingsJson)
-    {
-        var settingsPath = WriteSettingsFile(settingsJson);
-        var folder = Path.GetDirectoryName(Path.GetDirectoryName(settingsPath))!;
-
-        return new WorkspaceFolder
-        {
-            Name = name,
-            DocumentUri = ProtocolConversions.CreateAbsoluteDocumentUri(folder),
-        };
-    }
-
-    private static void CreateSolutionFile(WorkspaceFolder folder, string relativePath)
-    {
-        var folderPath = ProtocolConversions.GetDocumentFilePathFromUri(folder.DocumentUri.GetRequiredParsedUri());
-        var solutionPath = Path.Combine(folderPath, relativePath);
-        Directory.CreateDirectory(Path.GetDirectoryName(solutionPath)!);
-        File.WriteAllText(solutionPath, string.Empty);
     }
 }
