+++ src/LanguageServer/Microsoft.CodeAnalysis.LanguageServer/HostWorkspace/AutoLoadProjectsInitializer.cs
@@ -170,6 +170,6 @@ internal sealed class AutoLoadProjectsInitializer(
             }
         }
 
-        return default;
+        return (isLoadingDisabled: false, solutionPath: null);
     }
 }
